requirements.txt
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-snowflake==3.0.0